feat: 优化代码

+++ app/src/main/java/io/legado/app/data/dao/BookDao.kt
@@ -82,7 +82,7 @@ interface BookDao {
         bookUrl: String,
         durChapterIndex: Int,
         durChapterPos: Int,
-        durChapterTime: Int,
+        durChapterTime: Long,
         durChapterTitle: String?
     )
 }
